+++ tests/src/builder/mod.rs
@@ -1,3 +1,4 @@
 mod custom_builder;
 mod generic_tree;
+mod loc_info;
 mod use_context;