+++ tests/src/glr/special/right_nullable/mod.rs
@@ -1,4 +1,4 @@
-use rustemo::{rustemo_mod, GssHead, Parser, TreeBuilder};
+use rustemo::{rustemo_mod, Parser, TreeBuilder};
 use rustemo_compiler::output_cmp;
 
 rustemo_mod!(lang, "/src/glr/special/right_nullable");
@@ -16,14 +16,10 @@ fn glr_special_right_nullable_g2() {
             &format!("src/glr/special/right_nullable/tree_{}.ast", i),
             format!(
                 "{:#?}",
-                tree.unwrap().build::<TreeBuilder<
-                    '_,
-                    str,
-                    lang::ProdKind,
-                    lang::TokenKind,
-                >, GssHead<'_, str, lang::State, lang::TokenKind>, lang::State>(
-                    &mut builder
-                )
+                tree.unwrap()
+                    .build::<TreeBuilder<'_, str, lang::ProdKind, lang::TokenKind>, lang::State>(
+                        &mut builder
+                    )
             )
         );
     })